TXEvaluation/pyproject.toml
[project]
name = "tx7709evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-openai>=0.3.18",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
]
